load_files_api_drf/pyproject.toml
[tool.poetry]
name = "load-files-api-drf"
version = "0.1.0"
description = ""
authors = ["vova2092 <vovai2092@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
psycopg2 = "^2.9.7"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
celery = "^5.3.4"
redis = "^5.0.1"
pillow = "^10.0.1"


[tool.poetry.group.dev.dependencies]
coverage = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
